feat(evaluation): rehash project modules
