+++ crates/pollux/src/main.rs
@@ -7,7 +7,6 @@ mod ioc;
 mod pollux;
 
 use crate::infra::cli;
-use console::style;
 use tikv_jemallocator::Jemalloc;
 
 #[global_allocator]
@@ -15,43 +14,15 @@ static GLOBAL: Jemalloc = Jemalloc;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    better_panic::install();
-    human_panic::setup_panic!();
+    cli::troubleshooting::setup_troubleshooting();
 
-    env_logger::builder()
-        .format_timestamp(None)
-        .format_module_path(false)
-        .format_level(false)
-        .format_file(false)
-        .format_target(false)
-        .init();
-
-    let task = cli::parse_arguments()?;
+    let task = cli::parsing::parse_arguments()?;
     let pollux = ioc::create_pollux(task);
 
     println!("Evaluating veracity for packages. This operation may take some time ...");
 
     let results = pollux.execute().await?;
-    let statistics = results.statistics;
 
-    println!();
-    println!("Packages evaluated : {}", statistics.total_project_packages);
-    println!("Missing veracity factors : {}", statistics.without_veracity_level);
-    println!("With existing factors : {}", statistics.with_veracity_level);
-    println!();
-
-    results
-        .outcomes
-        .iter()
-        .for_each(|(package, maybe_veracity_check)| match maybe_veracity_check {
-            Some(level) => {
-                println!("For {} : veracity = {:?} ", package, style(level).cyan());
-            },
-            None => {
-                println!("For {} : {}", package, style("failed to evaluate").red());
-            },
-        });
-
-    println!();
+    cli::feedback::show_user_feedback(&results);
     Ok(())
 }
